+++ src/test/kotlin/com/codingame/game/AllTowersPlayer.kt
@@ -13,13 +13,13 @@ class AllTowersPlayer(stdin: InputStream, stdout: PrintStream, stderr: PrintStre
         // if touching a tower that isn't at max health, keep growing it
         val growingTower = obstacles
           .filter { it.owner == 0 && it.structureType == 1 && it.incomeRateOrHealthOrCooldown < 400 }
-          .firstOrNull { it.location.distanceTo(queenLoc) - it.radius - QUEEN_RADIUS < 5 }
+          .firstOrNull { it.location.distanceTo(queenLoc).toDouble - it.radius - QUEEN_RADIUS < 5 }
 
         if (growingTower != null) return "BUILD ${growingTower.obstacleId} TOWER"
 
         val queenTarget = obstacles
           .filter { it.owner == -1 }
-          .minBy { it.location.distanceTo(queenLoc) - it.radius } ?: return "WAIT"
+          .minBy { it.location.distanceTo(queenLoc).toDouble - it.radius } ?: return "WAIT"
 
         return "BUILD ${queenTarget.obstacleId} TOWER"
       }
