+++ src/test/kotlin/com/codingame/game/CSJPlayer.kt
@@ -37,13 +37,13 @@ class CSJPlayer(stdin: InputStream, stdout: PrintStream, stderr: PrintStream): B
         // if touching a tower that isn't at max health, keep growing it
         val growingTower = obstacles
           .filter { it.owner == 0 && it.structureType == 1 && it.incomeRateOrHealthOrCooldown < 400 }
-          .firstOrNull { it.location.distanceTo(queenLoc) - it.radius - QUEEN_RADIUS < 5 }
+          .firstOrNull { it.location.distanceTo(queenLoc).toDouble - it.radius - QUEEN_RADIUS < 5 }
         if (growingTower != null) return "BUILD ${growingTower.obstacleId} TOWER"
 
         // if touching a mine that isn't at max capacity, keep growing it
         val growingMine = obstacles
             .filter { it.owner == 0 && it.structureType == 0 && it.incomeRateOrHealthOrCooldown < it.maxResourceRate }
-            .firstOrNull { it.location.distanceTo(queenLoc) - it.radius - QUEEN_RADIUS < 5 }
+            .firstOrNull { it.location.distanceTo(queenLoc).toDouble - it.radius - QUEEN_RADIUS < 5 }
         if (growingMine != null) {
           return "BUILD ${growingMine.obstacleId} MINE"
         }
@@ -52,19 +52,19 @@ class CSJPlayer(stdin: InputStream, stdout: PrintStream, stderr: PrintStream): B
           .filter { it.owner == -1 || (it.owner == 1 && it.structureType != 1) }
           .filter { target -> !obstacles.any {
             it.owner == 1 && it.structureType == 1 &&
-              it.location.distanceTo(target.location) - it.attackRadiusOrCreepType - target.radius < -30 }}
-          .minBy { it.location.distanceTo(queenLoc) - it.radius }
+              it.location.distanceTo(target.location).toDouble - it.attackRadiusOrCreepType - target.radius < -30 }}
+          .minBy { it.location.distanceTo(queenLoc).toDouble - it.radius }
 
         if (queenTarget == null) {
           // bear toward closest friendly tower
           val closestTower = obstacles
             .filter { it.owner == 0 && it.structureType == 1 }
-            .minBy { it.location.distanceTo(queenLoc) - it.radius }
+            .minBy { it.location.distanceTo(queenLoc).toDouble - it.radius }
 
           return closestTower?.let { "BUILD ${it.obstacleId} TOWER" } ?: "WAIT"
         }
 
-        val dist = queenTarget.location.distanceTo(queenLoc) - QUEEN_RADIUS - queenTarget.radius
+        val dist = queenTarget.location.distanceTo(queenLoc).toDouble - QUEEN_RADIUS - queenTarget.radius
 
         if (dist < 5) {
           // Touching an obstacle; do something here
